+++ src/encoded/static/components/experiment.js
@@ -15,6 +15,7 @@ var pipeline = require('./pipeline');
 var reference = require('./reference');
 var biosample = require('./biosample');
 var software = require('./software');
+var sortTable = require('./sorttable');
 
 var Breadcrumbs = navbar.Breadcrumbs;
 var DbxrefList = dbxref.DbxrefList;
@@ -32,6 +33,8 @@ var PubReferenceList = reference.PubReferenceList;
 var ExperimentTable = dataset.ExperimentTable;
 var SingleTreatment = biosample.SingleTreatment;
 var SoftwareVersionList = software.SoftwareVersionList;
+var SortTablePanel = sortTable.SortTablePanel;
+var SortTable = sortTable.SortTable;
 
 
 var anisogenicValues = [
@@ -49,7 +52,7 @@ function humanFileSize(size) {
 }
 
 
-var testTableConfig = {
+var processTableConfig = {
     title: 'Processed data',
     columns: {
         'accession': {
@@ -238,6 +241,17 @@ var Experiment = module.exports.Experiment = React.createClass({
 
         var experiments_url = '/search/?type=experiment&possible_controls.accession=' + context.accession;
 
+        // Extract three kinds of file arrays
+        var files = _(this.props.items).groupBy(function(file) {
+            if (file.output_category === 'raw data') {
+                return 'raw';
+            } else if (file.output_category === 'reference') {
+                return 'ref';
+            } else {
+                return 'proc';
+            }
+        });
+
         // XXX This makes no sense.
         //var control = context.possible_controls[0];
         return (
@@ -420,7 +434,9 @@ var Experiment = module.exports.Experiment = React.createClass({
                 {context.files.length ?
                     <div>
                         <h3>Files linked to {context.accession}</h3>
-                        <FileTable items={context.files} encodevers={encodevers} anisogenic={anisogenic} />
+                        <SortTablePanel>
+                            <SortTable list={files.proc} config={processTableConfig} meta={{encodevers: encodevers, anisogenic: anisogenic}} />
+                        </SortTablePanel>
                     </div>
                 : null }
 
@@ -1482,114 +1498,3 @@ var QcDetailsView = function(metrics) {
         return null;
     }
 };
-
-
-var TableSortablePanel = React.createClass({
-    render: function() {
-        return (
-            <div className="table-panel table-file">
-                <div className="table-responsive">
-                    {this.props.children}
-                </div>
-            </div>
-        );
-    }
-});
-
-
-var TableSortable = React.createClass({
-    propTypes: {
-        list: React.PropTypes.array.isRequired, // Array of objects to display in the table
-        config: React.PropTypes.object.isRequired, // Defines the columns of the table
-        sortColumn: React.PropTypes.string // ID of column to sort by default; first column if not given
-    },
-
-    getInitialState: function() {
-        var sortColumn;
-
-        // Get the given sort column ID, or the default (first key in columns object) if none given
-        if (this.props.sortColumn) {
-            sortColumn = this.props.sortColumn;
-        } else {
-            sortColumn = Object.keys(this.props.config.columns)[0];
-        }
-
-        return {
-            sortColumn: sortColumn, // ID of currently sorting column
-            reversed: false // True if sorting of current sort column is reversed
-        };
-    },
-
-    // Handle clicks in the column headers for sorting columns
-    sortDir: function(column) {
-        var reversed = column === this.state.sortColumn ? !this.state.reversed : false;
-        this.setState({sortColumn: column, reversed: reversed});
-    },
-
-    // Called when any column needs sorting. If the column's config has a sorter function, call it
-    // to handle its sorting. Otherwise assume the values can be retrieved from the currently sorted column ID.
-    sortColumn: function(a, b) {
-        var result;
-        var columnId = this.state.sortColumn;
-        var sorter = this.props.config.columns[columnId].sorter;
-        if (sorter) {
-            result = sorter(a, b);
-        } else {
-            if (a[columnId] && b[columnId]) {
-                result = a[columnId] > b[columnId] ? 1 : -1;
-            } else {
-                result = a[columnId] ? -1 : (b[columnId] ? 1 : 0);
-            }
-        }
-        return this.state.reversed ? -result : result;
-    },
-
-    render: function() {
-        var list = this.props.list;
-        var config = this.props.config;
-        var columns = config.columns;
-        var columnIds = Object.keys(columns);
-        var colCount = columnIds.length;
-
-        return (
-            <table className="table table-striped">
-
-                <thead>
-                    <tr className="table-section"><th colSpan={colCount}>{config.title}</th></tr>
-                    <tr>
-                        {columnIds.map(columnId => {
-                            var columnClass = columnId === this.state.sortColumn ? (this.state.reversed ? 'tcell-desc' : 'tcell-asc') : 'tcell-sort';
-
-                            return (
-                                <th key={columnId} className="tcell-sortable" onClick={this.sortDir.bind(null, columnId)}>
-                                    <span>{columns[columnId].title}<i className={columnClass}></i></span>
-                                </th>
-                            );
-                        })}
-                    </tr>
-                </thead>
-
-                <tbody>
-                    {list.sort(this.sortColumn).map(item => {
-                        return (
-                            <tr key={item.uuid}>
-                                {columnIds.map(columnId => {
-                                    if (columns[columnId].display) {
-                                        return <td key={columnId}>{columns[columnId].display(item)}</td>;
-                                    }
-
-                                    // No custom display function; just display the standard way
-                                    var itemValue = columns[columnId].getValue ? columns[columnId].getValue(item) : item[columnId];
-                                    return (
-                                        <td key={columnId}>{itemValue}</td>
-                                    );
-                                })}
-                            </tr>
-                        );
-                    })}
-                </tbody>
-
-            </table>
-        );
-    }
-});
